+++ views/base.njk
@@ -9,5 +9,8 @@
     <div class="container">{% block contingut %} {% endblock %}</div>
   </body>
 
-  <footer>Pablo Cortés Bravo</footer>
+  <footer>
+    <p>Exercici Movies - Desenrrollament Web en Entorn Servidor</p>
+    <p>Pablo Cortés Bravo</p>
+    </footer>
 </html>
