+++ views/error.njk
@@ -1,19 +1,8 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <title>Error</title>
-        <link rel="stylesheet" href="/css/bootstrap.min.css">
-    </head>
-    <body>
-        <div class="container">
-            <h1>Error</h1>
-            <div class="alert alert-danger">
-                {% if error %}
-                    {{ error }}
-                {% else %}
-                    Error en l'aplicació
-                {% endif %}
-            </div>
-        </div>
-    </body>
-</html>
+{%extends 'base.njk'%}
+{%block titol%}Pel·lícules | Error {%endblock%}
+{%block contingut%}
+<div class="container">
+  <h1>Error</h1>
+  <div class="alert alert-danger" role="alert">{{ error }}</div>
+</div>
+{%endblock%}
