+++ views/movies/update.njk
@@ -1,0 +1,44 @@
+{% extends "../base.njk" %}
+
+{% block titol %}Pel·lícules | Editar{% endblock %}
+
+{% block contingut %}
+
+ <h1>Editar pel·lícula</h1>
+
+ <form action="/movies/update/{{movie._id}}" method="post">
+    <div class="form-group mb-3">
+        <label>Títol:</label>
+        <input type="text" class="form-control" name="title" placeholder="Títol de la pel·lícula..." value="{{ movie.title}}">
+    </div>
+    <div class="form-group mb-3">
+        <label>Any:</label>
+        <input type="number" class="form-control" name="year"
+                placeholder="Any de publicació..." value="{{ movie.year }}">
+        {%if errors.year %} <div class="text-danger"> {{errors.year.message}}  </div> {% endif %}
+    </div>
+    <div class="form-group mb-3">
+        <label>Director:</label>
+        <input type="text" class="form-control" name="director"
+                placeholder="Nom del director..." value="{{ movie.director }}">
+    </div>
+    <div class="form-group mb-3">
+        <label>Duració:</label>
+        <input type="number" class="form-control" name="duration"
+                placeholder="Duració en minuts..." value="{{ movie.duration }}">
+    </div>
+    <div class="form-group mb-3">
+        <label>Gènere:</label>
+        <select class="form-control" name="genre" multiple>
+            <option value="Sci-Fi" {% if 'Sci-Fi' in movie.genre %} selected {% endif %} >Sci-Fi</option>
+            <option value="Crime" {% if 'Crime' in movie.genre %} selected {% endif %} >Crime</option>
+            <option value="Drama" {% if 'Drama' in movie.genre %} selected {% endif %} >Drama</option>
+        </select>
+    </div>
+
+    <button type="submit" class="btn btn-primary">
+        Enviar
+    </button>
+ </form>
+
+{% endblock %}
